+++ include/render/primitives/Box.h
@@ -38,6 +38,10 @@ namespace render {
                                             const PrimitivePacketState4& states) const override;
     PrimitivePacketHit8 intersectPacketHits(const Ray8& rays,
                                             const PrimitivePacketState8& states) const override;
+    PrimitivePacketInterval4
+    intersectPacketIntervals(const Ray4& rays, const PrimitivePacketState4& states) const override;
+    PrimitivePacketInterval8
+    intersectPacketIntervals(const Ray8& rays, const PrimitivePacketState8& states) const override;
     virtual bool intersects(const Rayd& ray, render::State& state) const override;
 
     /**
@@ -76,6 +80,8 @@ namespace render {
   private:
     template<typename Packet, typename StateArray, typename Result>
     Result intersectPacketHitsFor(const Packet& rays, const StateArray& states) const;
+    template<typename Packet, typename StateArray, typename Result>
+    Result intersectPacketIntervalsFor(const Packet& rays, const StateArray& states) const;
 
     Vector3d m_center, m_edge;
   };